// _archive/backup_before_restoration_20251101_160150/testing/types/testing.ts
export interface Vulnerability {
  id: string;
  name: string;
  description: string;
  severity: 'critical' | 'high' | 'medium' | 'low' | 'info';
  cweid: number;            // Common Weakness Enumeration ID
  wascid?: number;          // Web Application Security Consortium ID
  url: string;
  method?: string;
  evidence?: string;
  solution: string;
  reference: string[];
  instances: VulnerabilityInstance[];
  owaspCategory?: string;
}

export interface VulnerabilityInstance {
  uri: string;
  method: string;
  param?: string;
  attack?: string;
  evidence?: string;
}